WIP Switch to apache commons instead of nio for directory monitor
